In controllers/userController.js, in exports.updateMe, console log both req.file which multer puts in the request body in the form of req.file; also console log req.body
POSTMAN: Login a user, select form-data and place name and a changed name along with photo and select file type and upload a photo.  This photo will be stored in /public/img/users in the form of a long number.  We will not have the ability to open this, but it will create something like this URL:  public/imgusers/2ab774e6d3121bfb599d347d4807829d