In controllers/userController.js, in exports.resizeUserPhoto, surround the function with catchAsync, turn the function into an async function and await call to sharp
NOTE: At this point the app only works on POSTMAN when requesting Update Tour route as an admin or lead-guide and inputing form-data on that request; While the images have not yet gone anywhere, because they have yet to be processed, you can update the text with Body set to form-data and field price: 997, for example. We also put form-data in for imageCover (uploading one photo) and images (uploading 3 photos).  These images do not display but any change in text does change.

// controllers/userController

// TODO: Resize user uploaded images
// Note: npm install sharp
exports.resizeUserPhoto = catchAsync(async (req, res, next) => {
  // If the user has not uploaded an image, go on to next middleware
  if (!req.file) return next();

  req.file.filename = `user-${req.user.id}-${Date.now()}.jpeg`;

  await sharp(req.file.buffer)
    .resize(500, 500)
    .toFormat('jpeg')
    .jpeg({ quality: 90 })
    .toFile(`public/img/users/${req.file.filename}`);

  next();
});